fix link resolve and ref group use in subset

+++ src/lib/preproc.hpp
@@ -42,14 +42,14 @@ public:
            double max = std::numeric_limits<double>::quiet_NaN(),
            double depth = std::numeric_limits<double>::quiet_NaN(),
            Interpolation interp = Interpolation::LINEAR,
-           double scale = 1.0,
+           float scale = 1.0,
            int every = 1
           );
   ACC_D(int, every, 1)
   ACC_D(double, min, std::numeric_limits<double>::quiet_NaN())
   ACC_D(double, max, std::numeric_limits<double>::quiet_NaN())
   ACC_D(double, depth, std::numeric_limits<double>::quiet_NaN())
-  ACC_D(double, scale, 1.0)
+  ACC_D(float, scale, 1.0)
   ACC_D(Interpolation, interpolation, Interpolation::LINEAR)
   ACC(cpath, intrinsics)
